ENH: move back Quantile regression tests
